backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-dotenv
pydantic-settings
